Tests/LibWeb/Text/expected/iframe-mouseout-when-leaving-nested-navigable.txt

child mouseout count: 1
